La till Level7

+++ src/model/levels/Level2.java
@@ -17,7 +17,7 @@ public class Level2 extends Level {
     }
 
     @Override
-    public int generateMathQuestion() {
+    public double generateMathQuestion() {
         int a, b;
         a = (int)(Math.random() * 10 + 1); // generera ett tal från 1 - 10
         b = (int)(Math.random() * 10 + 1); // generera ett tal från 1 - 10
